Fix #7793: Do not expose internal input API
Slow-Path input uses UINT16 for scancodes on wire, but only the
lower byte is actually used. (the extended fields are sent in
keyboardFlags field)
Hide this implementation detail and adjust the API to use UINT8
for the code instead just like the corresponding Fast-Path PDU

+++ libfreerdp/core/fastpath.c
@@ -674,7 +674,7 @@ static BOOL fastpath_recv_input_event_scancode(rdpFastPath* fastpath, wStream* s
 {
 	rdpInput* input;
 	UINT16 flags;
-	UINT16 code;
+	UINT8 code;
 
 	if (!fastpath || !fastpath->rdp || !fastpath->rdp->input || !s)
 		return FALSE;
